fix(migraciones): Correcciones de idempotencia en migraciones

+++ database/migrations/2025_04_25_154805_create_modules_table.php
@@ -8,6 +8,8 @@ return new class extends Migration
 {
     public function up(): void
     {
+        if (Schema::hasTable('modules')) return;
+
         Schema::create('modules', function (Blueprint $table) {
             $table->id();
             $table->foreignId('program_id')->constrained()->onDelete('cascade')->comment('id_programa externo');
