+++ database/migrations/2025_04_25_154918_create_module_classes_table.php
@@ -8,6 +8,8 @@ return new class extends Migration
 {
     public function up(): void
     {
+        if (Schema::hasTable('module_classes')) return;
+
         Schema::create('module_classes', function (Blueprint $table) {
             $table->id();
             $table->foreignId('module_id')->constrained()->onDelete('cascade');
