+++ database/migrations/2025_04_25_165521_create_attendances_table.php
@@ -11,6 +11,8 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('attendances')) return;
+
         Schema::create('attendances', function (Blueprint $table) {
             $table->id();
             $table->foreignId('module_class_id')->constrained('module_classes')->onDelete('cascade');
