+++ database/migrations/2025_05_13_183726_create_documents_table.php
@@ -11,6 +11,8 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('documents')) return;
+
         Schema::create('documents', function (Blueprint $table) {
             $table->id();
             $table->foreignId('inscription_id')->constrained()->onDelete('cascade');
