+++ database/migrations/2025_05_22_163517_create_document_followups_table.php
@@ -11,6 +11,8 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('document_followups')) return;
+
         Schema::create('document_followups', function (Blueprint $table) {
             $table->id();
             $table->foreignId('inscription_id')->constrained()->onDelete('cascade');
