+++ database/migrations/2025_10_02_172052_create_payment_requests_table.php
@@ -11,6 +11,8 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('payment_requests')) return;
+
         Schema::create('payment_requests', function (Blueprint $table) {
             $table->id();
             $table->foreignId('module_id')->constrained('modules')->onDelete('cascade');
